Quality of life update: separated models and conrollers into their own files
