+++ app/controllers/get-categories-controller.js
@@ -1,0 +1,7 @@
+const { fetchCategories } = require("../models");
+
+exports.getCategories = (request, response) => {
+  fetchCategories().then((categories) => {
+    response.status(200).send({ categories });
+  });
+};
