+++ app/controllers/index.js
@@ -1,0 +1,4 @@
+const { getCategories } = require("./get-categories-controller");
+const { getReview } = require("./get-review-controller");
+
+module.exports = { getCategories, getReview };
