+++ app/models/fetch-categories-model.js
@@ -1,0 +1,7 @@
+const db = require("../../db/connection");
+
+exports.fetchCategories = () => {
+  return db.query(`SELECT * FROM categories`).then(({ rows }) => {
+    return rows;
+  });
+};
