+++ app/models/index.js
@@ -1,0 +1,4 @@
+const { fetchCategories } = require("./fetch-categories-model");
+const { fetchReview } = require("./fetch-review-model");
+
+module.exports = { fetchCategories, fetchReview };
